login FE

+++ frontend/src/app/_services/token-storage.service.ts
@@ -1,4 +1,6 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { SITE_CONFIG } from '../app.config';
+import { isPlatformBrowser } from '@angular/common';
 
 const TOKEN_KEY = 'auth-token';
 const USER_KEY = 'auth-user';
@@ -7,32 +9,50 @@ const USER_KEY = 'auth-user';
   providedIn: 'root'
 })
 export class TokenStorageService {
-  constructor() { }
+  constructor(
+    @Inject(PLATFORM_ID) private platformId: Object,
+  ) { }
 
   signOut(): void {
-    window.sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.clear();
+    }
   }
 
   public saveToken(token: string): void {
-    window.sessionStorage.removeItem(TOKEN_KEY);
-    window.sessionStorage.setItem(TOKEN_KEY, token);
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.removeItem(TOKEN_KEY);
+      window.sessionStorage.setItem(TOKEN_KEY, token);
+    }
   }
 
   public getToken(): string | null {
-    return window.sessionStorage.getItem(TOKEN_KEY);
+    if (isPlatformBrowser(this.platformId)) {
+      return window.sessionStorage.getItem(TOKEN_KEY);
+    }
+    else {
+      return null
+    }
   }
 
   public saveUser(user: any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.removeItem(USER_KEY);
+      let userEnc = (SITE_CONFIG.settings.security.encodeSessionValues) ? window.btoa(JSON.stringify(user)) : JSON.stringify(user);
+      window.sessionStorage.setItem(USER_KEY, userEnc);
+    }
   }
 
   public getUser(): any {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return JSON.parse(user);
-    }
+    if (isPlatformBrowser(this.platformId)) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      if (user) {
+        let userDec= (SITE_CONFIG.settings.security.encodeSessionValues) ? JSON.parse(window.atob(user)) : JSON.parse(user);
+        // console.log(userDec)
+        return userDec
+      }
 
-    return {};
+      return {};
+    }
   }
 }
